Support Secure/Insexure eventstore
