+++ VoucherManagement.BusinessLogic.Tests/VoucherManagement.BusinessLogic.Tests.csproj
@@ -10,7 +10,7 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0.5" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.2.0" />
     <PackageReference Include="Moq" Version="4.18.1" />
-    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.3" />
+    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.7" />
     <PackageReference Include="Shouldly" Version="4.0.3" />
     <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="17.0.15" />
     <PackageReference Include="xunit" Version="2.4.1" />
