+++ VoucherManagement.BusinessLogic/VoucherManagement.BusinessLogic.csproj
@@ -5,15 +5,15 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="EstateManagement.Client" Version="1.1.1" />
-    <PackageReference Include="EstateReporting.Database" Version="1.1.2" />
+    <PackageReference Include="EstateManagement.Client" Version="1.1.2" />
+    <PackageReference Include="EstateReporting.Database" Version="1.1.3" />
     <PackageReference Include="MediatR" Version="10.0.1" />
-    <PackageReference Include="MessagingService.Client" Version="1.1.0" />
+    <PackageReference Include="MessagingService.Client" Version="1.1.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.5" />
     <PackageReference Include="NetBarcode" Version="1.5.0" />
-    <PackageReference Include="SecurityService.Client" Version="1.1.0" />
-    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.3" />
-    <PackageReference Include="Shared.EventStore" Version="1.3.3" />
+    <PackageReference Include="SecurityService.Client" Version="1.1.1" />
+    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.7" />
+    <PackageReference Include="Shared.EventStore" Version="1.3.7" />
     <PackageReference Include="System.IO.Abstractions" Version="17.0.15" />
 	  <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="6.0.5" />
   </ItemGroup>
