+++ VoucherManagement.Client/VoucherManagement.Client.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ClientProxyBase" Version="1.3.3" />
+    <PackageReference Include="ClientProxyBase" Version="1.3.7" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
