+++ VoucherManagement.IntegrationTests/VoucherManagement.IntegrationTests.csproj
@@ -6,17 +6,15 @@
 
   <ItemGroup>
 	  <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="6.0.5" />
-	  <PackageReference Include="EstateManagement.Client" Version="1.1.1" />
-    <PackageReference Include="SecurityService.Client" Version="1.1.0" />
-    <PackageReference Include="Shared" Version="1.3.3" />
-    <PackageReference Include="Shared.IntegrationTesting" Version="1.3.3" />
+	  <PackageReference Include="EstateManagement.Client" Version="1.1.2" />
+    <PackageReference Include="SecurityService.Client" Version="1.1.1" />
+    <PackageReference Include="Shared" Version="1.3.7" />
+    <PackageReference Include="Shared.IntegrationTesting" Version="1.3.7" />
     <PackageReference Include="Shouldly" Version="4.0.3" />
     <PackageReference Include="SpecFlow" Version="3.9.74" />
     <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.9.74" />
     <PackageReference Include="SpecFlow.xUnit" Version="3.9.74" />
-    <PackageReference Include="EventStore.Client.Grpc.PersistentSubscriptions" Version="22.0.0" />
-    <PackageReference Include="EventStore.Client.Grpc.ProjectionManagement" Version="22.0.0" />
-    <PackageReference Include="EstateReporting.Database" Version="1.1.2" />
+    <PackageReference Include="EstateReporting.Database" Version="1.1.3" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
       <PrivateAssets>all</PrivateAssets>
