+++ VoucherManagement.Voucher.DomainEvents/VoucherManagement.Voucher.DomainEvents.csproj
@@ -1,11 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
 	  <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.3" />
+    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.7" />
   </ItemGroup>
 
 </Project>
