+++ VoucherManagement.VoucherAggregate/VoucherManagement.VoucherAggregate.csproj
@@ -1,11 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
 	  <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Shared.EventStore" Version="1.3.3" />
+    <PackageReference Include="Grpc.Net.Client" Version="2.47.0" />
+    <PackageReference Include="Shared.EventStore" Version="1.3.7" />
   </ItemGroup>
 
   <ItemGroup>
