+++ VoucherManagement/Startup.cs
@@ -142,12 +142,10 @@ namespace VoucherManagement
         public void ConfigureContainer(ServiceRegistry services)
         {
             ConfigurationReader.Initialise(Startup.Configuration);
-
-            Startup.ConfigureEventStoreSettings();
-
-            services.IncludeRegistry<MiddlewareRegistry>();
+            
             services.IncludeRegistry<MediatorRegistry>();
             services.IncludeRegistry<RepositoryRegistry>();
+            services.IncludeRegistry<MiddlewareRegistry>();
             services.IncludeRegistry<ClientRegistry>();
             services.IncludeRegistry<DomainEventHandlerRegistry>();
             services.IncludeRegistry<DomainServiceRegistry>();
@@ -170,28 +168,12 @@ namespace VoucherManagement
         /// Configures the event store settings.
         /// </summary>
         /// <param name="settings">The settings.</param>
-        internal static void ConfigureEventStoreSettings(EventStoreClientSettings settings = null)
+        public static void ConfigureEventStoreSettings(EventStoreClientSettings settings)
         {
-            if (settings == null)
-            {
-                settings = new EventStoreClientSettings();
-            }
-
-            settings.CreateHttpMessageHandler = () => new SocketsHttpHandler
-                                                      {
-                                                          SslOptions =
-                                                          {
-                                                              RemoteCertificateValidationCallback = (sender,
-                                                                                                     certificate,
-                                                                                                     chain,
-                                                                                                     errors) => true,
-                                                          }
-                                                      };
-
-            settings.ConnectionName = Startup.Configuration.GetValue<String>("EventStoreSettings:ConnectionName");
             settings.ConnectivitySettings = EventStoreClientConnectivitySettings.Default;
             settings.ConnectivitySettings.Address = new Uri(Startup.Configuration.GetValue<String>("EventStoreSettings:ConnectionString"));
             settings.ConnectivitySettings.Insecure = Startup.Configuration.GetValue<Boolean>("EventStoreSettings:Insecure");
+
 
             settings.DefaultCredentials = new UserCredentials(Startup.Configuration.GetValue<String>("EventStoreSettings:UserName"),
                                                               Startup.Configuration.GetValue<String>("EventStoreSettings:Password"));
@@ -204,6 +186,19 @@ namespace VoucherManagement
     public static class Extensions
     {
         #region Fields
+
+        public static IServiceCollection AddInSecureEventStoreClient(
+            this IServiceCollection services,
+            Uri address,
+            Func<HttpMessageHandler>? createHttpMessageHandler = null)
+        {
+            return services.AddEventStoreClient((Action<EventStoreClientSettings>)(options =>
+                                                                                   {
+                                                                                       options.ConnectivitySettings.Address = address;
+                                                                                       options.ConnectivitySettings.Insecure = true;
+                                                                                       options.CreateHttpMessageHandler = createHttpMessageHandler;
+                                                                                   }));
+        }
 
         private static readonly Action<TraceEventType, String, String> log = (tt,
                                                                               subType,
