+++ VoucherManagement/VoucherManagement.csproj
@@ -17,8 +17,6 @@
 		<PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="6.0.4" />
 		<PackageReference Include="AspNetCore.HealthChecks.UI.InMemory.Storage" Version="6.0.4" />
 		<PackageReference Include="AspNetCore.HealthChecks.Uris" Version="6.0.3" />
-		<PackageReference Include="EventStore.Client.Grpc.PersistentSubscriptions" Version="22.0.0" />
-		<PackageReference Include="EventStore.Client.Grpc.ProjectionManagement" Version="22.0.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.5" />
 		<PackageReference Include="MediatR" Version="10.0.1" />
 		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.5" />
@@ -33,8 +31,8 @@
 		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="6.0.5" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
 		<PackageReference Include="NLog.Extensions.Logging" Version="5.0.0" />
-		<PackageReference Include="Shared" Version="1.3.3" />
-		<PackageReference Include="Shared.EventStore" Version="1.3.3" />
+		<PackageReference Include="Shared" Version="1.3.7" />
+		<PackageReference Include="Shared.EventStore" Version="1.3.7" />
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.1" />
 		<PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.3.1" />
 		<PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="7.0.3" />
